+++ lib/automake.mk
@@ -284,6 +284,7 @@ lib_libopenvswitch_la_SOURCES = \
 	lib/vconn-stream.c \
 	lib/vconn.c \
 	lib/versions.h \
+	lib/vl-mff-map.h \
 	lib/vlan-bitmap.c \
 	lib/vlan-bitmap.h \
 	lib/vlog.c \
