+++ lib/nx-match.h
@@ -34,6 +34,7 @@ struct ofpact_stack;
 struct ofpbuf;
 struct nx_action_reg_load;
 struct nx_action_reg_move;
+struct vl_mff_map;
 
 
 /* Nicira Extended Match (NXM) flexible flow match helper functions.
